REF: collect+parametrize reorder_levels tests (#32373)
